define EC2ConsoleClient subclass of txaws EC2Client
(per davidsarah's suggestion). The subclass is necessary because the
query action submitted to AWS is 'GetConsoleOutput', there is no
method in the txaws.ec2.client.EC2Client class for this action.
Since this action is necessary for side channel fingerprint
verification of the server ssh pubkey, it might be worthwhile to
submit a patch to txaws.  Also defined are a parser for the response
and helpers.  hostpubkeyextractor seems particularly fragile to me.