verify SSEC2 sshfp on sign up, and store in known hosts
Since there are two channels by which we communicate with AWS, and since one is potentially
authenticatable we:

Obtain the new (deployed on signup) SSEC2s ssh pubkey fp over https, and verify that subsequent
ssh sessions are shown the same fingerprint.  We than store the associated pubkey in
/home/website/.ssh/knownhosts . This means that subsequent communication (e.g. fabric deploys)
has a degree of cryptographic authenticity.