implements statmover monitoring on an EC2
This includes a specific 'EmitEvent' for the 'rss' of a process
using Zooko's 'get_mem_used'.  To add new EmitEvents a name is
defined and added to the eventemissions_config.json file.  The
generate script must then define appropriate behavior to:

(1)  Acquire the data e.g. get_mem_used
(2)  Format the data e.g. make_emit_string/structure
(3)  emit the data e.g. the check_call to emit-client

NOTE:  I wrote this some time ago.  We now have the convention of using FilePath
so it'd be nice to refactor file IO to use it soon.