Optimizes signup flow to remove reboot during call to install_server.
This optimization cuts off _at least_ 4 minutes of wait from a signup.
This optimization was designed and tested with:  ami-10314d79
The relevant products (s4) should be changed in lae_automation_config.json
to deploy "ami-10314d79" and the change should be deployed to secret_config
concurrently with this optimization_commit.