Add functionality for handling stripe messages
NOTE: submission of data from the User's
browser to stripe is handled by the previously
committed "subscription_signup.js".  Unittests
for that code are also in an ancestral commit.

Note: non-solvent payment credentials, cause stripe servers
to send a message to the browser.
Note: invalid submission data, e.g. wrong type of CC, is
handled in browser by our html/and-or stripe.js.

This handler is invoked when:
(1) The User's payment (cc) information has been
processed by stripe.
(2) That information is valid
    i.  valid-and-solvent number, CVC, exp date
(3) Stripe has responded to the requested processing.

Handler functionality:
1. The handler loads our secret stripe api key, and
2. attempts to charge the User's cc (referenced by a unique stripe token), for
a subscription, to our "S4" plan at stripe.

3a. handles several possible exceptions generated by stripe.
3b. In the case of a confirmation from stripe that subscription succeeded,
i.e. 'the card was charged', the handler:
   I.  logs a successful subscription on the webserver fs
   II. invokes the full_signup service provide by the flappserver service

4. In all cases renders an appropriate webpage and returns it to the
twisted web-server for subsequent transmission to the User's browser.

Also included are incomplete unittests for the above functionality.