Define the flappservice "full_signup" stripe style
This module/process/service defines the interface
the automation flappserver provides to the
webserver.

Once the webserver has a successful subscription
acknowledged by stripe, it submits the relevant
information (via FlappCommand.run) to the process
defined in this module.

After creating and/or appending to several log files
some unique to each signup, some of which persist across
subscriptions, this function initiates the provisioning
process with a call to "activate_subscribed_service".