Compare subscription ids like-to-like
The subscription id that is a label is base32 encoded.
The subscription id that is an annotation is unencoded.
The comparisons are being made against unencoded subscription
identifiers.

Comparing the encoded form always results in no match found.  This
caused the convergence loop to delete all ReplicaSets and Pods.  The
result was behavior where these resources were perpetually created and
deleted.

Comparing against the unencoded form now properly leaves ReplicaSets and
Pods for active subscriptions alone, the desired behavior.

The test suite failed to catch this problem because the assertions for
ReplicaSets and Pods allowed them to be missing (in case they simply had
not yet been created).  The test suite has been improved so that as soon
as they are created, they become required by the check phase.  This lets
us assert exactly the correct ReplicaSet and Pod state.