Toss old Eliot logs
Since Hypothesis runs through this code many times for a "single" test,
quite a large amount of Eliot logs accumulate before the end of the
test.  Most of the logs are from older test runs and can be very verbose
individually (because Hypothesis may have arranged to create 10
subscriptions or some such), let alone all in a big pile.

Since Hypothesis should be working towards a maximally shrunk
demonstration of the problem, the final run should have the most
succinct logs (as it has the most succinct inputs to trigger the
problem).  So keeping just the last set of logs should be ideal.