Expand the window even further
With one new sample every 15m, a 15m window is basically not a running
average at all.  The benchmark results are pretty noisy for some
reason (to be determined).  Expand the window to cover several samples
to smooth out the graph at least a little bit.

It's not clear at this time what resolution would be most useful on
these graphs.  This change is meant to make the graphs a little more
visually simple to parse.  It may turn out that this is a misleading
simplicity.  If so then we'll do something else.