Update liveness on progress
When performance is very bad, it's hard to complete even one successful
measurement within a reasonable liveness period.  So make any progress
sufficient to mark the process as alive.  This seems reasonable since if
we manage any reads or writes we must be making some kind of progress.

There is a separate "are we actually completing benchmarks" metric to be
concerned about.  This could be added on top of the progress work.